ml_model/requirements.txt
# Diabetic Retinopathy Detection — Python Dependencies
# Install: pip install -r requirements.txt

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0          # EfficientNet-B4 via timm

# Image processing
opencv-python>=4.7.0
Pillow>=9.0.0
albumentations>=1.3.0

# Data & ML
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web API
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
tqdm>=4.65.0

# MongoDB
pymongo>=4.4.0
python-dotenv>=1.0.0
